feat: add cobra cmd's
